+++ src/com.dol/ui/elements/ContextMenu.java
@@ -16,6 +16,7 @@ public class ContextMenu extends JPopupMenu {
         JMenuItem itemSpeed = new JMenuItem("Change air balloon speed");
         JMenuItem itemSetPoint = new JMenuItem("Set point A");
 
+        //TODO: добавить вызов метода полёта шара из бэкенда
         itemStart.addActionListener(e -> pressedOnStart = true);
         itemSpeed.addActionListener(e -> new SpeedControllerWindow(frame).setVisible(true));
 
@@ -25,7 +26,7 @@ public class ContextMenu extends JPopupMenu {
                 itemSetPoint.setText("Set point B");
             } else if (pressedOnA) {
 
-                if (!imageObserver.getPointA().isWorth()) {
+                if (imageObserver.getPointA().isEmpty()) {
                     JOptionPane.showMessageDialog(this, "Select point A", "Warning!", JOptionPane.WARNING_MESSAGE);
                     return;
                 }
@@ -35,7 +36,7 @@ public class ContextMenu extends JPopupMenu {
                 itemSetPoint.setText("Done");
             } else {
 
-                if (!imageObserver.getPointB().isWorth()) {
+                if (imageObserver.getPointB().isEmpty()) {
                     JOptionPane.showMessageDialog(this, "Select point B", "Warning!", JOptionPane.WARNING_MESSAGE);
                     return;
                 }
