--- src/com/dol/ui/exceptions/FileNotFoundException.java
@@ -1,12 +1,0 @@
-package com.dol.ui.exceptions;
-
-public class FileNotFoundException extends RuntimeException {
-
-    public FileNotFoundException() {
-        super();
-    }
-
-    public FileNotFoundException(String err) {
-        super(err);
-    }
-}
